smartagent[extension/receiver]: Handle all defaults in extension
