+++ internal/extension/smartagentextension/config_test.go
@@ -52,16 +52,17 @@ func TestLoadConfig(t *testing.T) {
 			},
 			bundleDir: bundleDir,
 			collectdConfig: config.CollectdConfig{
-				Timeout:             40,
-				ReadThreads:         5,
-				WriteThreads:        2,
-				WriteQueueLimitHigh: 500000,
-				WriteQueueLimitLow:  400000,
-				LogLevel:            "notice",
-				IntervalSeconds:     0,
-				WriteServerIPAddr:   "127.9.8.7",
-				WriteServerPort:     0,
-				ConfigDir:           "/var/run/signalfx-agent/collectd",
+				Timeout:              40,
+				ReadThreads:          5,
+				WriteThreads:         2,
+				WriteQueueLimitHigh:  500000,
+				WriteQueueLimitLow:   400000,
+				LogLevel:             "notice",
+				IntervalSeconds:      10,
+				WriteServerIPAddr:    "127.9.8.7",
+				WriteServerPort:      0,
+				ConfigDir:            "/var/run/signalfx-agent/collectd",
+				HasGenericJMXMonitor: true,
 			},
 		}
 	}(), emptyConfig)
@@ -77,16 +78,18 @@ func TestLoadConfig(t *testing.T) {
 			},
 			bundleDir: "/opt/bin/collectd/",
 			collectdConfig: config.CollectdConfig{
-				Timeout:             10,
-				ReadThreads:         1,
-				WriteThreads:        4,
-				WriteQueueLimitHigh: 5,
-				WriteQueueLimitLow:  1,
-				LogLevel:            "info",
-				IntervalSeconds:     5,
-				WriteServerIPAddr:   "10.100.12.1",
-				WriteServerPort:     9090,
-				ConfigDir:           "/etc/",
+				Timeout:              10,
+				ReadThreads:          1,
+				WriteThreads:         4,
+				WriteQueueLimitHigh:  5,
+				WriteQueueLimitLow:   1,
+				LogLevel:             "info",
+				IntervalSeconds:      5,
+				WriteServerIPAddr:    "10.100.12.1",
+				WriteServerPort:      9090,
+				ConfigDir:            "/etc/",
+				BundleDir:            "/opt/bin/collectd/",
+				HasGenericJMXMonitor: true,
 			},
 		}
 	}(), allSettingsConfig)
@@ -100,18 +103,20 @@ func TestLoadConfig(t *testing.T) {
 				TypeVal: "smartagent",
 				NameVal: "smartagent/partial_settings",
 			},
-			bundleDir: bundleDir,
+			bundleDir: "/opt/",
 			collectdConfig: config.CollectdConfig{
-				Timeout:             10,
-				ReadThreads:         1,
-				WriteThreads:        4,
-				WriteQueueLimitHigh: 5,
-				WriteQueueLimitLow:  400000,
-				LogLevel:            "notice",
-				IntervalSeconds:     0,
-				WriteServerIPAddr:   "127.9.8.7",
-				WriteServerPort:     0,
-				ConfigDir:           "/etc/",
+				Timeout:              10,
+				ReadThreads:          1,
+				WriteThreads:         4,
+				WriteQueueLimitHigh:  5,
+				WriteQueueLimitLow:   400000,
+				LogLevel:             "notice",
+				IntervalSeconds:      10,
+				WriteServerIPAddr:    "127.9.8.7",
+				WriteServerPort:      0,
+				BundleDir:            "/opt/",
+				ConfigDir:            "/etc/",
+				HasGenericJMXMonitor: true,
 			},
 		}
 	}(), partialSettingsConfig)
@@ -140,16 +145,18 @@ func TestSmartAgentConfigProvider(t *testing.T) {
 
 	require.Equal(t, func() config.CollectdConfig {
 		return config.CollectdConfig{
-			Timeout:             10,
-			ReadThreads:         1,
-			WriteThreads:        4,
-			WriteQueueLimitHigh: 5,
-			WriteQueueLimitLow:  1,
-			LogLevel:            "info",
-			IntervalSeconds:     5,
-			WriteServerIPAddr:   "10.100.12.1",
-			WriteServerPort:     9090,
-			ConfigDir:           "/etc/",
+			Timeout:              10,
+			ReadThreads:          1,
+			WriteThreads:         4,
+			WriteQueueLimitHigh:  5,
+			WriteQueueLimitLow:   1,
+			LogLevel:             "info",
+			IntervalSeconds:      5,
+			WriteServerIPAddr:    "10.100.12.1",
+			WriteServerPort:      9090,
+			BundleDir:            "/opt/bin/collectd/",
+			ConfigDir:            "/etc/",
+			HasGenericJMXMonitor: true,
 		}
 	}(), saConfigProvider.CollectdConfig())
 	require.Equal(t, "/opt/bin/collectd/", saConfigProvider.BundleDir())
