+++ internal/extension/smartagentextension/testdata/config.yaml
@@ -14,6 +14,7 @@ extensions:
       writeServerPort: 9090
       configDir: /etc/
   smartagent/partial_settings:
+    bundleDir: /opt/
     collectd:
       timeout: 10
       readThreads: 1
